Removed HelperService, HelperServiceImpl, and HelperServiceTest and redistributed its functions
