+++ src/main/java/edu/hawaii/its/api/service/GroupAttributeService.java
@@ -3,7 +3,10 @@ package edu.hawaii.its.api.service;
 import edu.hawaii.its.api.type.Grouping;
 import edu.hawaii.its.api.type.GroupingsServiceResult;
 import edu.hawaii.its.api.type.OptRequest;
+import edu.hawaii.its.api.type.Person;
 import edu.hawaii.its.api.type.SyncDestination;
+
+import edu.internet2.middleware.grouperClient.ws.beans.ResultMetadataHolder;
 
 import java.util.List;
 
@@ -23,5 +26,12 @@ public interface GroupAttributeService {
     GroupingsServiceResult assignGrouperPrivilege(String privilegeName, String groupPath, boolean isSet);
 
     GroupingsServiceResult updateDescription(String groupPath, String ownerUsername, String description);
+
+    GroupingsServiceResult makeGroupingsServiceResult(ResultMetadataHolder resultMetadataHolder, String action);
+
+    GroupingsServiceResult makeGroupingsServiceResult(ResultMetadataHolder resultMetadataHolder, String action,
+                                                      Person person);
+
+    GroupingsServiceResult makeGroupingsServiceResult(String resultCode, String action);
 
 }
