+++ src/test/java/edu/hawaii/its/api/service/TestGrouperApiService.java
@@ -95,9 +95,6 @@ public class TestGrouperApiService {
     @Autowired
     MembershipService membershipService;
 
-    @Autowired
-    HelperService helperService;
-
     @BeforeAll
     public void init() {
         assertTrue(memberAttributeService.isAdmin(ADMIN));
