+++ src/test/java/edu/hawaii/its/api/service/TestGrouperApiServiceTutorial.java
@@ -113,9 +113,6 @@ public class TestGrouperApiServiceTutorial {
     @Autowired
     MembershipService membershipService;
 
-    @Autowired
-    HelperService helperService;
-
     @BeforeAll
     public void init() {
         assertTrue(memberAttributeService.isAdmin(ADMIN));
